move PackageManager to config.rs
